requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==3.3.9
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.44.0
colorama==0.4.6
dill==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-environ==0.12.0
django-filter==25.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
jiter==0.9.0
mccabe==0.7.0
openai==1.77.0
packaging==25.0
pillow==11.1.0
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==6.30.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pylint==3.3.5
pyparsing==3.2.3
python-dotenv==1.0.1
python-http-client==3.3.7
redis==5.2.1
requests==2.32.3
rsa==4.9
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tomlkit==0.13.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
